[ARC-584] Rearrange packages
make command implementations sub to cli parsing

branch: feature/ARC-584-discover-service-dependencies
