--- src/mpyl/cli/projects/__init__.py
@@ -1,1 +1,0 @@
-"""CLI commands related to projects"""
